Explicit `modifier` on cards to group changes better
This will significantly improve readability, and make both
code & data easier to understand.

This will add more fields to data-files, but since these will be automated
anyway, it won't hurt if they are easy to understand when
editing manually until then.

And if this format adds too much redundant weight for production
environment, we could find a optimized data format for
the prod version of the game.